+++ lib/helpers/helper.dart
@@ -730,3 +730,8 @@ event<T>({Map? data}) async =>
 /// ```
 dump(dynamic value, {String? tag, bool alwaysPrint = false}) =>
     NyLogger.dump(value, tag, alwaysPrint: alwaysPrint);
+
+/// Sleep for a given amount of milliseconds.
+sleep(int milliseconds) async {
+  await Future.delayed(Duration(milliseconds: milliseconds));
+}